LPS-146923 Move Sidebar components to his own folder
