+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/components/Layout/Layout.tsx
@@ -15,8 +15,8 @@
 import {Outlet} from 'react-router-dom';
 
 import HeaderContextProvider from '../../context/HeaderContext';
+import Sidebar from '../Sidebar';
 import Header from './Header';
-import Sidebar from './Sidebar';
 
 const Layout = () => (
 	<main className="testray-main">
