+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/components/Sidebar/index.tsx
